changing position of button
