+++ src/app/body/team-details/label-card/label-card.component.html
@@ -12,7 +12,7 @@
 * See the MIT License for more details. 
 ***********************************************************/ -->
 <div class="card" style="width: 18rem;" *ngIf="labelsReady">
-    <div class="card-header">{{scope}}    
+    <div class="card-header">{{scope}} 
     <mat-icon fontSet="material-icons" style="cursor: pointer;" (click)="addLabel()" data-toggle="modal" data-target="#addLabel">
         add
     </mat-icon></div>
