+++ src/app/body/view-organization-details/teams-card/teams-card.component.html
@@ -39,7 +39,6 @@
         </div>
         <div class="row">
             <div class="col">
-                <h3 (click)="createDefaultLabels()">Add Default Label</h3>
                 <h2 class="card-title" (click)="openTeamDetails()" style="cursor: pointer;">{{ team.TeamName }}</h2>
                 <h3 class="card-subtitle text-muted">{{ team.TeamDescription }}</h3>
             </div>
